Add DiscreteUniform test to test_step.py.
