+++ pymc3/tests/test_step.py
@@ -37,11 +37,13 @@ from pymc3.distributions import (
     Beta,
     HalfNormal,
 )
+from pymc3.distributions.discrete import DiscreteUniform, Poisson
 
 from numpy.testing import assert_array_almost_equal
 import numpy as np
 import numpy.testing as npt
 import pytest
+from unittest import TestCase
 import theano
 import theano.tensor as tt
 from .helpers import select_by_precision
@@ -1047,3 +1049,24 @@ class TestNutsCheckTrace:
             ]
         )
         assert (trace.model_logp == model_logp_).all()
+
+class TestDiscreteUniform(TestCase):
+
+    def test_DiscreteUniform(self, lower=0, upper=200000000, obs=5000000, draws=20000):
+        """Test change allowing CategoricalGibbsMetropolis step method with a DiscreteUniform distribution.
+        Test checks that MCMC runs OK."""
+        obs = theano.shared(obs)
+
+        with Model() as model2:
+            x = DiscreteUniform('x', lower, upper - 1)
+            sfs_obs = Poisson('sfs_obs', mu=x, observed=obs)
+
+        with model2:
+            step = CategoricalGibbsMetropolis([x])
+            trace = sample(draws, tune=0, step=step)
+        return trace
+
+    def test_bad_lower(self):
+        """Check that lower !=0 raises an error."""
+        with pytest.raises(ValueError):
+            self.test_DiscreteUniform(lower=1, upper=200000001, obs=5000000, draws=20000)
